+++ webapp/python/isucari/app.py
@@ -10,7 +10,7 @@ from sqlalchemy.exc import SQLAlchemyError
 from . import (utils,
                database,
                models,
-               isucari,
+               core,
                )
 from .config import Constants, Config, base_dir, static_folder
 from .cache import cache
@@ -74,9 +74,9 @@ def post_initialize():
     subprocess.call([d + "/init.sh"])
 
     payment_service_url = flask.request.json.get('payment_service_url', Constants.DEFAULT_PAYMENT_SERVICE_URL)
-    isucari.save_config('payment_service_url', payment_service_url)
+    core.save_config('payment_service_url', payment_service_url)
     shipment_service_url = flask.request.json.get('shipment_service_url', Constants.DEFAULT_SHIPMENT_SERVICE_URL)
-    isucari.save_config('shipment_service_url', shipment_service_url)
+    core.save_config('shipment_service_url', shipment_service_url)
 
     return flask.jsonify({
         "campaign": 0,  # キャンペーン実施時には還元率の設定を返す。詳しくはマニュアルを参照のこと。
@@ -94,7 +94,7 @@ def get_new_items():
     if created_at < 0:
         raise HttpException(requests.codes.bad_request, "created_at param error")
 
-    items = isucari.timeline(item_id, created_at)
+    items = core.timeline(item_id, created_at)
     has_next = False
     if len(items) > Constants.ITEMS_PER_PAGE:
         has_next = True
@@ -116,7 +116,7 @@ def get_new_category_items(root_category_id=None):
     if created_at < 0:
         raise HttpException(requests.codes.bad_request, "created_at param error")
 
-    root_category, items = isucari.category_items(item_id, created_at, root_category_id)
+    root_category, items = core.category_items(item_id, created_at, root_category_id)
 
     has_next = False
     if len(items) > Constants.ITEMS_PER_PAGE:
@@ -143,7 +143,7 @@ def get_transactions():
     if created_at < 0:
         raise HttpException(requests.codes.bad_request, "created_at param error")
 
-    items = isucari.transaction_items(user, item_id, created_at)
+    items = core.transaction_items(user, item_id, created_at)
 
     has_next = False
     if len(items) > Constants.TRANSACTIONS_PER_PAGE:
@@ -168,7 +168,7 @@ def get_user_items(user_id=None):
     if created_at < 0:
         raise HttpException(requests.codes.bad_request, "created_at param error")
 
-    items = isucari.user_time_line(user, item_id, created_at)
+    items = core.user_time_line(user, item_id, created_at)
     has_next = False
     if len(items) > Constants.ITEMS_PER_PAGE:
         has_next = True
@@ -184,7 +184,7 @@ def get_user_items(user_id=None):
 @app.route("/items/<int:item_id>.json", methods=["GET"])
 def get_item(item_id=None):
     user = get_current_user()
-    item = isucari.get_item(user, item_id)
+    item = core.get_item(user, item_id)
 
     return flask.jsonify(item.for_detail_json())
 
@@ -198,7 +198,7 @@ def post_item_edit():
     item_id = int(data['item_id'])
     user = get_current_user()
 
-    item = isucari.edit(user, item_id, price)
+    item = core.edit(user, item_id, price)
 
     return flask.jsonify(dict(
         item_id=item.id,
@@ -216,7 +216,7 @@ def post_buy():
     item_id = int(flask.request.json['item_id'])
     token = flask.request.json['token']
 
-    transaction_evidence_id = isucari.buy(buyer, item_id, token)
+    transaction_evidence_id = core.buy(buyer, item_id, token)
 
     return flask.jsonify(dict(transaction_evidence_id=transaction_evidence_id))
 
@@ -233,13 +233,13 @@ def post_sell():
     if "image" not in flask.request.files:
         raise HttpException(requests.codes.internal_server_error, 'image error')
 
-    item = isucari.sell(get_current_user(),
-                        flask.request.form['name'],
-                        flask.request.form['description'],
-                        int(flask.request.form['price']),
-                        flask.request.form['category_id'],
-                        flask.request.files['image']
-                        )
+    item = core.sell(get_current_user(),
+                     flask.request.form['name'],
+                     flask.request.form['description'],
+                     int(flask.request.form['price']),
+                     flask.request.form['category_id'],
+                     flask.request.files['image']
+                     )
 
     return flask.jsonify({
         'id': item.id,
@@ -252,7 +252,7 @@ def post_ship():
     user = get_current_user()
     item_id = flask.request.json["item_id"]
 
-    transaction_id, reserve_id = isucari.ship(user, item_id)
+    transaction_id, reserve_id = core.ship(user, item_id)
     return flask.jsonify(dict(
         path="/transactions/{}.png".format(transaction_id),
         reserve_id=reserve_id,
@@ -265,7 +265,7 @@ def post_ship_done():
     user = get_current_user()
     item_id = flask.request.json["item_id"]
 
-    transaction_evidence = isucari.ship_done(user, int(item_id))
+    transaction_evidence = core.ship_done(user, int(item_id))
 
     return flask.jsonify(dict(transaction_evidence_id=transaction_evidence.id))
 
@@ -276,7 +276,7 @@ def post_complete():
     user = get_current_user()
     item_id = flask.request.json["item_id"]
 
-    transaction_evidence = isucari.complete(user, item_id)
+    transaction_evidence = core.complete(user, item_id)
 
     return flask.jsonify(dict(transaction_evidence_id=transaction_evidence.id))
 
@@ -287,7 +287,7 @@ def get_qrcode(transaction_evidence_id):
         raise HttpException(requests.codes.bad_request, "incorrect transaction_evidence id")
 
     seller = get_current_user()
-    shipping = isucari.get_qr_code(seller, transaction_evidence_id)
+    shipping = core.get_qr_code(seller, transaction_evidence_id)
 
     res = flask.make_response(shipping.img_binary)
     res.headers.set('Content-Type', 'image/png')
@@ -301,7 +301,7 @@ def post_bump():
     data = ensure_required_payload(['item_id'])
     user = get_current_user()
 
-    item = isucari.bump(user, data)
+    item = core.bump(user, data)
 
     return flask.jsonify({
         'item_id': item.id,
@@ -320,7 +320,7 @@ def get_settings():
     outputs['csrf_token'] = flask.session.get('csrf_token', '')
     categories = models.Category.query.all()
     outputs['categories'] = [c.for_json() for c in categories]
-    outputs['payment_service_url'] = isucari.get_payment_service_url()
+    outputs['payment_service_url'] = core.get_payment_service_url()
 
     return flask.jsonify(outputs)
 
@@ -328,7 +328,7 @@ def get_settings():
 @app.route("/login", methods=["POST"])
 def post_login():
     data = ensure_required_payload(['account_name', 'password'])
-    user = isucari.login(data)
+    user = core.login(data)
 
     flask.session['user_id'] = user.id
     flask.session['csrf_token'] = utils.random_string(10)
@@ -340,7 +340,7 @@ def post_login():
 @app.route("/register", methods=["POST"])
 def post_register():
     data = ensure_required_payload(['account_name', 'password', 'address'])
-    user = isucari.register(data)
+    user = core.register(data)
 
     flask.session['user_id'] = user.id
     flask.session['csrf_token'] = utils.random_string(10)
